update testbench template generator for new changes

+++ testbench-template-generator.py
@@ -13,17 +13,13 @@ https://github.com/COMP10001/py-safe-grading-framework
 # Choose from these options to generate a testbench file template
 ASSIGNMENT_DESC = "Assignment 1 Testbench"
 TESTCASE_AUTHOR = ""
-INCLUDE_PEP8_CHECK = False
-INCLUDE_AST_CHECK = False
-NUM_VISIBLE_FUNCTION_TESTS = 1
-NUM_VISIBLE_SCRIPT_TESTS = 0
-NUM_HIDDEN_FUNCTION_TESTS = 1
-NUM_HIDDEN_SCRIPT_TESTS = 0
-NUM_PRIVATE_FUNCTION_TESTS = 0
-NUM_PRIVATE_SCRIPT_TESTS = 0
+INCLUDE_PEP8_CHECK = True
+INCLUDE_AST_CHECK = True
+NUM_FUNCTION_TESTS = 1
+NUM_SCRIPT_TESTS = 1
 STUDENT_FUNCTION = "test"
 STUDENT_FILE_NAME = "program.py"
-STUDENT_FILE_PATH_PREFIX = "/home/"
+SHOW_TEST_REPORTS = False
 
 
 ############################################################################################
@@ -41,67 +37,75 @@ License: MIT
 The latest version and documentation can be found at:
 https://github.com/COMP10001/py-safe-grading-framework
 """
-from safetestingframework import *
+from pysafegradingfw import *
 
 ### ENSURE PER TESTCASE SCORING IS ENABLED!
 ### ENSURE TIMEOUT (s) IS SET HIGHER THAN THE SUM OF 'function_timeout_seconds' FOR EACH TEST!
 ### SEE TESTBENCH EXAMPLE FOR MORE EXPLANATION
 
-# Remove the test file after loading by Ed, to prevent ability to print out contents
-os.remove(__file__)
+# Remove the test file after loading by Ed, to prevent ability to print out content of testcases.
+# In your local dev environment run with `python testbench.py --local-dev` to prevent deletion!
+if not (len(sys.argv) >= 2 and sys.argv[1] == "--local-dev"):
+    os.remove(__file__)
 
 # DANGER: Overrides Hidden and Private Tests so they become visible,
-# Need to re-run test cases for all students for it to work
+# Need to re-run test cases for all students for it to work.
 RELEASE_TEST_CASES = False
 
 STUDENT_FUNCTION = "{2}"
 STUDENT_FILE_NAME = "{3}"
-STUDENT_FILE_PATH_PREFIX = "{4}"
 
-# In setup mode all testcases will fail to ensure it is not left enabled
-# Given input and output data will be formatted as a string with no line wrapping for easy copy pasting into testbench
-SETUP_MODE = False
+test_bench = SafeTesting(
+    debug_mode = False,
+    format_test_in_out_data_as_str = False,
+    make_all_tests_visible = False,
+    show_all_passed_tests_first = True,
+    show_test_reports = {4},
+    # file_path_prefix
+)
 
-SHOW_ALL_PASSED_TESTS_FIRST = True
-
+# Any files required for hidden tests should be removed
+# from path except during that specific test
 FILES_TO_HIDE = [] # eg ["abc.txt"]
-HIDDEN_FILE_DICT = cache_hidden_test_files(FILES_TO_HIDE)
+test_bench.cache_hidden_test_files(FILES_TO_HIDE)
+'''.format(ASSIGNMENT_DESC, TESTCASE_AUTHOR, STUDENT_FUNCTION, STUDENT_FILE_NAME, SHOW_TEST_REPORTS)
 
-class SafeTesting():'''.format(ASSIGNMENT_DESC, TESTCASE_AUTHOR, STUDENT_FUNCTION, STUDENT_FILE_NAME, STUDENT_FILE_PATH_PREFIX)
 
 SUFFIX = \
 r'''
 if __name__ == "__main__":
-    run_tests(SafeTesting, setup_mode=SETUP_MODE, show_all_passed_tests_first=SHOW_ALL_PASSED_TESTS_FIRST)
+    test_bench.run_tests()
 '''
 
 TEST_PEP8 = \
 r'''
-    @setname() # Set the student visible testcase name as the function name excluding `test` or `test_` with spaces instead of '_'
-    @score(0)
-    def test_PEP8_Check(self):
-        PEP8_IGNORED = 'E121,E123,E125,E126,E127,E128,E129,E221,E222,E223,E224,E225,E131,E133,E301,E302,E303,E304,E731,F401,F403,W2,W3,W503'
-        run_pep8_test(
-            student_file_name=STUDENT_FILE_NAME,                 # File to test function from
-            student_file_path_prefix=STUDENT_FILE_PATH_PREFIX,   # File path prefix, could change by Ed, but otherwise does not need to be touched
-            ignored_tests=PEP8_IGNORED                           # Modify as desired, this is the default value set in the framework.
-        )
-'''
+test_bench.register_pep8_test(
+    name = "PEP8 Check",
+    score = 0,
+    hidden = False,
+    private = False,
+    student_file_name = STUDENT_FILE_NAME,
+    # ignored_tests,
+)'''
 
 TEST_ASTCHECK = \
 r'''
-    @setname()
-    @score(0)
-    def testAST_Check(self):
-        run_astcheck_test(
-            student_file_name=STUDENT_FILE_NAME,               # File to test function from
-            student_file_path_prefix=STUDENT_FILE_PATH_PREFIX, # File path prefix, could change by Ed, but otherwise does not need to be touched
-            required_nodes=[],                                 # Eg [ast.For, ast.While] as a list/tuple or with description eg {ast.For: "for loop"}, see run_astcheck_test
-            non_allowed_nodes=[],                              # Eg [ast.For, ast.While] as a list/tuple or with description eg {ast.For: "for loop"}, see run_astcheck_test
-            non_allowed_functions=[],                          # Function names of any specific functions to disallow
-            non_allowed_imports=[],                            # Imports that are not allowed anywhere in student file or any local imports
-        )
-'''
+# Note function/script tests can also have ast checks
+test_bench.register_ast_test(
+    name = "AST Check",
+    score = 0,
+    hidden = False,
+    private = False,
+    student_file_name = STUDENT_FILE_NAME,
+    non_allowed_nodes = [],
+    non_allowed_functions = [],
+    non_allowed_methods = [],
+    non_allowed_imports = [],
+    required_nodes = [],
+    required_functions = [],
+    required_methods = [],
+    required_imports = [],
+)'''
 
 RUN_FUNCTION_TEST = \
 r'''
@@ -124,8 +128,7 @@ test_bench.register_function_test(
     expected_stderr = "",
     expected_files = [],
     files_to_reveal = [],
-    # custom verification function/feedback injection available
-    # ast check options available
+    # see documentation on repo for all options
 )'''
 
 RUN_SCRIPT_TEST = \
@@ -143,30 +146,8 @@ test_bench.register_script_test(
     expected_stderr = "",
     expected_files = [],
     files_to_reveal = [],
-    # custom verification function/feedback injection available
-    # ast check options available
+    # see documentation on repo for all options
 )'''
-
-
-TEST_VISIBLE = \
-r'''
-    @setname()
-    @score(0)
-    def testVisible_{0}(self):'''
-
-TEST_HIDDEN = \
-r'''
-    @setname()
-    @hidden(RELEASE_TEST_CASES) # Set the test case as hidden, override to become visible by setting RELEASE_TEST_CASES = True
-    @score(0)
-    def testHidden_{0}(self):'''
-
-TEST_PRIVATE = \
-r'''
-    @setname()
-    @private(RELEASE_TEST_CASES) # Set the test case as private, override to become visible by setting RELEASE_TEST_CASES = True
-    @score(0)
-    def testPrivate_{0}(self):'''
 
 # .txt because on ed it wont allow access to the file unless it ends in txt
 with open("testbench.txt", "w") as fp:
@@ -176,28 +157,11 @@ with open("testbench.txt", "w") as fp:
     if INCLUDE_AST_CHECK:
         fp.write(TEST_ASTCHECK+ '\n')
 
-    for i in range(NUM_VISIBLE_FUNCTION_TESTS):
-        fp.write(TEST_VISIBLE.format(i+1))
+    for i in range(NUM_FUNCTION_TESTS):
         fp.write(RUN_FUNCTION_TEST + '\n')
 
-    for i in range(NUM_VISIBLE_SCRIPT_TESTS):
-        fp.write(TEST_VISIBLE.format(i+NUM_VISIBLE_FUNCTION_TESTS+1))
+    for i in range(NUM_SCRIPT_TESTS):
         fp.write(RUN_SCRIPT_TEST + '\n')
 
-    for i in range(NUM_HIDDEN_FUNCTION_TESTS):
-        fp.write(TEST_HIDDEN.format(i+1))
-        fp.write(RUN_FUNCTION_TEST + '\n')
-
-    for i in range(NUM_HIDDEN_SCRIPT_TESTS):
-        fp.write(TEST_HIDDEN.format(i+NUM_HIDDEN_FUNCTION_TESTS+1))
-        fp.write(RUN_SCRIPT_TEST + '\n')
-
-    for i in range(NUM_PRIVATE_FUNCTION_TESTS):
-        fp.write(TEST_PRIVATE.format(i+1))
-        fp.write(RUN_FUNCTION_TEST + '\n')
-
-    for i in range(NUM_PRIVATE_SCRIPT_TESTS):
-        fp.write(TEST_PRIVATE.format(i+NUM_PRIVATE_FUNCTION_TESTS+1))
-        fp.write(RUN_SCRIPT_TEST + '\n')
 
     fp.write(SUFFIX)
